test: harden chat transcript validation (#107893)
